+++ synalinks-server/inference-training/app/remote_llm/remote_llm_togetherai.py
@@ -5,15 +5,16 @@ import together
 
 class RemoteLLMTogetherAI(RemoteLLM):
     model: str = "togethercomputer/RedPajama-INCITE-7B-Instruct"
-   
+
     def generate(
-            self,
-            prompt: str,
-            temperature: float,
-            top_p: float,
-            max_tokens: int,
-            stop_list: List[str],
-        ) -> str:
+        self,
+        prompt: str,
+        temperature: float,
+        top_p: float,
+        max_tokens: int,
+        stop_list: List[str],
+    ) -> str:
+        together.api_key = self.api_key
         output = together.Complete.create(
             prompt=self.prompt,
             model=self.model,
@@ -26,15 +27,24 @@ class RemoteLLMTogetherAI(RemoteLLM):
         )
         return output.choices[0].text.strip()
 
-    def async_generate(
-            self,
-            prompt: str,
-            temperature: float,
-            top_p: float,
-            max_tokens: int,
-            stop_list: List[str],
-        ) -> str:
-        pass #TODO
+    async def async_generate(
+        self,
+        prompt: str,
+        max_tokens: int,
+        stop_list: List[str],
+    ) -> str:
+        together.api_key = self.api_key
+        output = together.Complete.create(
+            prompt=self.prompt,
+            model=self.model,
+            max_tokens=max_tokens,
+            temperature=0.7,
+            top_k=60,
+            top_p=0.9,
+            repetition_penalty=1.1,
+            stop=stop_list,
+        )
+        return output.choices[0].text.strip()
 
     def list_models(self) -> List[Any]:
         return together.Models.list()
